fix: don't be strict about nulls in relations
When relations have non-null values, the backend will sanitize those and
remove them. The reason why we should not force non-null values through
graphql, is because clients (like pza) which have a lot of other
services that insert data into mongo, might insert values with null
(sadly). This will then cause an error in the frontend when trying to
edit an entity, since the initial value is null and thus conflicts with
the graphql decleration.